Shifted Middlewares into seperate dedicated folder
