+++ index.js
@@ -6,8 +6,8 @@ const path = require("path")
 const bodyParser = require("body-parser")
 const jwt = require("jsonwebtoken")
 const bcrypt = require("bcrypt")
-const seller_auth = require("./middleware").seller_auth;
-const customer_auth = require("./middleware").customer_auth;
+const seller_auth = require("./middlewares/seller_auth");
+const customer_auth = require("./middlewares/customer_auth");
 const { response } = require("express");
 const app = express()
 const port = 5000;
