+++ middlewares/seller_auth.js
@@ -12,33 +12,7 @@ const client = new Client({
 client.connect();
 
 //Authenticating token from seller
-const seller_auth = async(req, res, next) => {
-    try {
-        const token = req.headers.authorization.split(" ")[1];
-        const decodedToken = jwt.verify(token, `${process.env.JWT_KEY}`);
-        const username = decodedToken.username;
-        req.locals = { 'seller_username': username }
-        await client
-            .query(`select * from sellers where username='${username}'`)
-            .then(response => {
-                if (response.rows.length == 0) {
-                    throw 'Invalid user ID';
-                } else {
-                    next();
-                }
-            })
-    } catch {
-        err => {
-            console.log(err)
-            res.status(401).json({
-                error: new Error('Invalid request!')
-            });
-        }
-    }
-
-};
-
-const customer_auth = async(req, res, next) => {
+module.exports = async(req, res, next) => {
     try {
         const token = req.headers.authorization.split(" ")[1];
         const decodedToken = jwt.verify(token, `${process.env.JWT_KEY}`);
@@ -63,8 +37,3 @@ const customer_auth = async(req, res, next) => {
     }
 
 };
-
-module.exports = {
-    'seller_auth': seller_auth,
-    'customer_auth': customer_auth,
-}
